Section 13: Authentication, Authorization & Sessions

+++ App/controllers/ListingController.php
@@ -2,9 +2,10 @@
 
 namespace App\Controllers;
 
-use Error;
 use Framework\Database;
+use Framework\Session;
 use Framework\Validation;
+use Framework\Authorization;
 
 class ListingController {
   protected $db;
@@ -16,7 +17,7 @@ class ListingController {
   }
   
   public function index() {
-    $listings = $this->db->query('SELECT * FROM listings')->fetchAll();
+    $listings = $this->db->query('SELECT * FROM listings ORDER BY created_at DESC')->fetchAll();
     loadView('listings/index', ['listings' => $listings]);
   }
 
@@ -57,7 +58,7 @@ class ListingController {
     ];
 
     $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
-    $newListingData['user_id'] = 1;
+    $newListingData['user_id'] = Session::get('user')['id'];
     $newListingData = array_map('sanitize', $newListingData);
 
     $requiredFields = [
@@ -99,6 +100,8 @@ class ListingController {
       $query = "INSERT INTO listings ({$fields}) VALUES ({$values})";
       $this->db->query($query, $newListingData);
 
+      Session::setFlashMessage('success_message', 'Listing created successfully');
+
       redirect('/listings');
       
     }
@@ -113,15 +116,22 @@ class ListingController {
 
     $listing = $this->db->query('SELECT * FROM listings WHERE id = :id', $params)->fetch();
 
+    // Check if listing exists
     if(!$listing){
       ErrorController::notFound('Listing not found');
       return;
     }
 
+    //Authorization
+    if(!Authorization::isOwner($listing->user_id)) {
+      Session::setFlashMessage('error_message', 'You are not authorized to delete this listing');
+      return redirect('/listings/' . $listing->id);
+    }
+
     $this->db->query('DELETE FROM listings WHERE id = :id', $params);
 
     // Set flash message
-    $_SESSION['success_message'] = 'Listing deleted successfully';
+    Session::setFlashMessage('success_message', 'Listing deleted successfully');
 
     redirect('/listings');
   }
@@ -140,6 +150,12 @@ class ListingController {
       return;
     }
 
+    //Authorization
+    if(!Authorization::isOwner($listing->user_id)) {
+      Session::setFlashMessage('error_message', 'You are not authorized to update this listing');
+      return redirect('/listings/' . $listing->id);
+    }
+
     loadView('listings/edit', ['listing' => $listing]);
   }
 
@@ -151,10 +167,17 @@ class ListingController {
     ];
 
     $listing = $this->db->query('SELECT * FROM listings WHERE id = :id', $params)->fetch();
+    
     // Check if listing exists
     if(!$listing) {
       ErrorController::notFound('Listing not found');
       return;
+    }
+
+    //Authorization
+    if(!Authorization::isOwner($listing->user_id)) {
+      Session::setFlashMessage('error_message', 'You are not authorized to update this listing');
+      return redirect('/listings/' . $listing->id);
     }
 
     $allowedFields = [
@@ -213,7 +236,8 @@ class ListingController {
       $updateValues['id'] = $id;
       $this->db->query($updateQuery, $updateValues);
 
-      $_SESSION['success_message'] = 'Listing Updated';
+      // Set flash message
+      Session::setFlashMessage('success_message', 'Listing updated successfully');
 
       redirect('/listings/' . $id);
     }
